[BUGFIX] Re-Enable RTE in Chrome 73
In Chrome 73 the "Cross-Origin Resource Policy" changed and seems
to affect loading external resources.

The RTE uses such an resource when it tries to display an interactive
HTML preview using an Iframe. In current Chrome versions this will cause
the preview to display unwanted markup instead of the desired preview.

Load the Iframe with an empty page instead of JavaScript to fix this issue.

Issue: https://github.com/FriendsOfTYPO3/rtehtmlarea/issues/26

TYPO3 ELTS: https://typo3.com/blog/typo3-7635-and-6241-elts-released